Forget to save newCosine as sine

+++ exponent.c
@@ -2,13 +2,13 @@
 #include <math.h>
 
 int main() {
-    double x, sum = 1;
+    long double x, sum = 1;
     double term = 1;
     double error = 0.00001;
     int i = 1;
 
     printf("Enter x: ");
-    scanf("%lf", &x);
+    scanf("%Lf", &x);
     
 
     while (fabs(term) > error) {
@@ -16,7 +16,10 @@ int main() {
         sum = sum + term;
         i++;
     }
-    printf("Exp of %lf is %lf.\n", x, sum);
+    printf("Exp of %1.0Lf is %1.0Lf.\n", x, sum);
     printf("Term is %lf.", term);
 }
 
+/* Breaks at x = 714
+ * I still don't understand why cosine breaks much faster, might be because of its alternating series
+ */
